# fix: Use vectorized search for username-password validation.