--- a/Projects/UOContent.Tests/Tests/Accounting/Security/PasswordProtectionTest.cs
+++ b/Projects/UOContent.Tests/Tests/Accounting/Security/PasswordProtectionTest.cs
@@ -17,30 +17,30 @@ public class PasswordProtectionTest
     [InlineData(typeof(HashAlgorithmPasswordProtection), "SHA2")]
     public void TestValidates(Type protectionType, string algorithmType)
     {
-            IPasswordProtection passwordProtection;
-            if (protectionType == typeof(HashAlgorithmPasswordProtection))
+        IPasswordProtection passwordProtection;
+        if (protectionType == typeof(HashAlgorithmPasswordProtection))
+        {
+            passwordProtection = algorithmType switch
             {
-                passwordProtection = algorithmType switch
-                {
-                    "SHA1" => HashAlgorithmPasswordProtection.SHA1Instance,
-                    "SHA2" => HashAlgorithmPasswordProtection.SHA2Instance,
-                    _      => HashAlgorithmPasswordProtection.MD5Instance,
-                };
-            }
-            else
-            {
-                passwordProtection = Activator.CreateInstance(protectionType) as IPasswordProtection;
-            }
-
-            if (passwordProtection == null)
-            {
-                Assert.Fail($"{protectionType.Name} is not an IPasswordProtection.");
-            }
-
-            var encryptedPassword = passwordProtection.EncryptPassword(plainPassword);
-
-            Assert.True(passwordProtection.ValidatePassword(encryptedPassword, plainPassword));
+                "SHA1" => HashAlgorithmPasswordProtection.SHA1Instance,
+                "SHA2" => HashAlgorithmPasswordProtection.SHA2Instance,
+                _      => HashAlgorithmPasswordProtection.MD5Instance,
+            };
         }
+        else
+        {
+            passwordProtection = Activator.CreateInstance(protectionType) as IPasswordProtection;
+        }
+
+        if (passwordProtection == null)
+        {
+            Assert.Fail($"{protectionType.Name} is not an IPasswordProtection.");
+        }
+
+        var encryptedPassword = passwordProtection.EncryptPassword(plainPassword);
+
+        Assert.True(passwordProtection.ValidatePassword(encryptedPassword, plainPassword));
+    }
 
     [Theory]
     [InlineData(typeof(Argon2PasswordProtection), null)]
@@ -50,28 +50,28 @@ public class PasswordProtectionTest
     [InlineData(typeof(HashAlgorithmPasswordProtection), "SHA2")]
     public void TestPasswordDoesNotValidate(Type protectionType, string algorithmType)
     {
-            IPasswordProtection passwordProtection;
-            if (protectionType == typeof(HashAlgorithmPasswordProtection))
+        IPasswordProtection passwordProtection;
+        if (protectionType == typeof(HashAlgorithmPasswordProtection))
+        {
+            passwordProtection = algorithmType switch
             {
-                passwordProtection = algorithmType switch
-                {
-                    "SHA1" => HashAlgorithmPasswordProtection.SHA1Instance,
-                    "SHA2" => HashAlgorithmPasswordProtection.SHA2Instance,
-                    _      => HashAlgorithmPasswordProtection.MD5Instance,
-                };
-            }
-            else
-            {
-                passwordProtection = Activator.CreateInstance(protectionType) as IPasswordProtection;
-            }
-
-            if (passwordProtection == null)
-            {
-                Assert.Fail($"{protectionType.Name} is not an IPasswordProtection.");
-            }
-
-            var encryptedPassword = passwordProtection.EncryptPassword(plainPassword);
-
-            Assert.False(passwordProtection.ValidatePassword(encryptedPassword, "Not the same password"));
+                "SHA1" => HashAlgorithmPasswordProtection.SHA1Instance,
+                "SHA2" => HashAlgorithmPasswordProtection.SHA2Instance,
+                _      => HashAlgorithmPasswordProtection.MD5Instance,
+            };
         }
+        else
+        {
+            passwordProtection = Activator.CreateInstance(protectionType) as IPasswordProtection;
+        }
+
+        if (passwordProtection == null)
+        {
+            Assert.Fail($"{protectionType.Name} is not an IPasswordProtection.");
+        }
+
+        var encryptedPassword = passwordProtection.EncryptPassword(plainPassword);
+
+        Assert.False(passwordProtection.ValidatePassword(encryptedPassword, "Not the same password"));
+    }
 }