--- a/Projects/UOContent/Accounting/AccountHandler.cs
+++ b/Projects/UOContent/Accounting/AccountHandler.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Buffers;
 using System.Collections.Generic;
 using System.Net;
 using ModernUO.CodeGeneratedEvents;
@@ -17,13 +18,13 @@ public static class AccountHandler
 
     private static int MaxAccountsPerIP;
     private static bool AutoAccountCreation;
-    private static bool RestrictDeletion = !TestCenter.Enabled;
-    private static TimeSpan DeleteDelay = TimeSpan.FromDays(7.0);
+    private static readonly bool RestrictDeletion = !TestCenter.Enabled;
+    private static readonly TimeSpan DeleteDelay = TimeSpan.FromDays(7.0);
     private static bool PasswordCommandEnabled;
 
     private static Dictionary<IPAddress, int> m_IPTable;
 
-    private static char[] m_ForbiddenChars = { '<', '>', ':', '"', '/', '\\', '|', '?', '*' };
+    private static readonly SearchValues<char> ForbiddenChars = SearchValues.Create("<>:\"/\\|?*");
 
     public static AccessLevel LockdownLevel { get; set; }
 
@@ -240,19 +241,6 @@ public static class AccountHandler
     public static bool CanCreate(IPAddress ip) =>
         !IPTable.TryGetValue(ip, out var result) || result < MaxAccountsPerIP;
 
-    private static bool IsForbiddenChar(char c)
-    {
-        for (var i = 0; i < m_ForbiddenChars.Length; ++i)
-        {
-            if (c == m_ForbiddenChars[i])
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
     private static Account CreateAccount(NetState state, string un, string pw)
     {
         if (un.Length == 0 || pw.Length == 0)
@@ -260,19 +248,16 @@ public static class AccountHandler
             return null;
         }
 
-        var isSafe = !(un.StartsWithOrdinal(" ") ||
-                       un.EndsWithOrdinal(" ") ||
-                       un.EndsWithOrdinal("."));
+        var unSpan = un.AsSpan();
+        var pwSpan = pw.AsSpan();
 
-        for (var i = 0; isSafe && i < un.Length; ++i)
-        {
-            isSafe = un[i] >= 0x20 && un[i] < 0x7F && !IsForbiddenChar(un[i]);
-        }
-
-        for (var i = 0; isSafe && i < pw.Length; ++i)
-        {
-            isSafe = pw[i] >= 0x20 && pw[i] < 0x7F;
-        }
+        // Usernames must not start with a space, end with a space, or end with a period
+        var isSafe = !(un.StartsWith(' ') || un.EndsWith(' ') || un.EndsWith('.')) &&
+                     // Usernames must only contain characters [0x20 -> 0x7E], and not contain any forbidden characters
+                     !unSpan.ContainsAnyExceptInRange((char)0x20, (char)0x7E) &&
+                     !unSpan.ContainsAny(ForbiddenChars) &&
+                     // Passwords must have characters [0x20 -> 0x7E]
+                     !pwSpan.ContainsAnyExceptInRange((char)0x20, (char)0x7E);
 
         if (!isSafe)
         {